projects/Integracion_Medilink_Pipedrive/cloud_functions/github/persona_update/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudevents==0.3.0
Flask==1.0.2
functions-framework==2.0.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.0
pandas==1.0.5
pathtools==0.1.2
pipedrive-python-lib==1.1.4
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
urllib3==1.25.9
watchdog==0.10.3
Werkzeug==1.0.1
